raspberryPie/requirements.txt
pip install --upgrade pip

opencv-python-headless
numpy
requests
python-dotenv
